Move to standalone paint signal file
