+++ src/OpenLoco/Objects/TrainSignalObject.h
@@ -27,7 +27,7 @@ namespace OpenLoco
     struct TrainSignalObject
     {
         string_id name;
-        uint16_t flags;       // 0x02
+        uint16_t flags;            // 0x02
         uint8_t animationSpeed;    // 0x04
         uint8_t num_frames;        // 0x05
         uint16_t cost_factor;      // 0x06
