+++ src/OpenLoco/Paint/PaintSignal.h
@@ -1,0 +1,12 @@
+#pragma once
+
+namespace OpenLoco::Map
+{
+    struct SignalElement;
+}
+namespace OpenLoco::Paint
+{
+    struct PaintSession;
+
+    void paintSignal(PaintSession& session, const Map::SignalElement& elSignal);
+}
